refactor: converts all imports to aliased and removes coupling in queries
