+++ src/config/icon.ts
