+++ src/config/query-keys.ts
@@ -1,0 +1,7 @@
+const queryKeys = {
+    news: 'GET_NEWS',
+    coins: 'GET_COINS',
+    coinPrices: 'GET_COIN_PRICES',
+};
+
+export default queryKeys;
