+++ src/features/coin-list/api/coins.ts
@@ -1,10 +1,11 @@
-import config from '../../../utils/config';
-import fetcher from '../../../utils/fetcher';
+import config from '@/config';
+import fetcher from '@/utils/fetcher';
+import { CoinDataResponse, CoinPriceResponse } from '../types';
 
 export function fetchCoinData() {
-    return fetcher('get', config.apiUrl, 'data/all/coinlist');
+    return fetcher<CoinDataResponse>('get', config.apiUrl, 'data/all/coinlist');
 }
 
 export function fetchCoinPrice(coinCodes: string[]) {
-    return fetcher('get', config.apiUrl, `data/pricemulti?fsyms=${coinCodes.join(',')}&tsyms=USD,GBP`);
+    return fetcher<CoinPriceResponse>('get', config.apiUrl, `data/pricemulti?fsyms=${coinCodes.join(',')}&tsyms=USD,GBP`);
 }
