+++ src/features/coin-list/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+
+export * from './routes/CoinList';
+export * from './routes/FullCoinList';
