--- src/features/coin-list/types/Coin.ts
@@ -1,6 +1,0 @@
-export interface Coin {
-    id: string;
-    code: string;
-    name: string;
-    imageURL: string;
-}
