--- src/features/coin-list/types/CoinPrice.ts
@@ -1,4 +1,0 @@
-export interface CoinPrice {
-    GBP: string;
-    USD: string;
-}
