+++ src/features/news/api/getNews.ts
@@ -1,4 +1,5 @@
-import config from '../../../utils/config';
-import fetcher from '../../../utils/fetcher';
+import config from '@/config';
+import fetcher from '@/utils/fetcher';
+import { NewsResult } from '../types';
 
-export const getNews = () => fetcher('get', config.apiUrl, 'data/v2/news/?lang=EN');
+export const getNews = () => fetcher<NewsResult>('get', config.apiUrl, 'data/v2/news/?lang=EN');
