--- src/features/news/types/NewsResult.ts
@@ -1,5 +1,0 @@
-import { NewsArticle } from './NewsArticle';
-
-export interface NewsResult {
-    Data: NewsArticle[];
-}
