+++ src/utils/divideDataIntoSubArrays.ts
@@ -9,7 +9,3 @@ export function divideDataIntoSubArrays<T>(data: T[], size = 50) {
         return acc;
     }, []);
 }
-
-export function getNumberRangeArray(start: number, end: number) {
-    return Array.from(Array(end - start + 1).keys()).map((key) => start + key);
-}
