+++ src/utils/getNumberRangeArray.ts
@@ -1,0 +1,3 @@
+export function getNumberRangeArray(start: number, end: number) {
+    return Array.from(Array(end - start + 1).keys()).map((key) => start + key);
+}
