+++ tsconfig.json
@@ -14,8 +14,12 @@
         "resolveJsonModule": true,
         "isolatedModules": true,
         "noEmit": true,
-        "jsx": "react-jsx"
+        "jsx": "react-jsx",
+        "baseUrl": ".",
+        "paths": {
+            "@/*": ["./src/*"]
+        }
     },
-    "include": ["src"],
+    "include": ["src", "vite.config.ts"],
     "references": [{ "path": "./tsconfig.node.json" }]
 }
