+++ vite.config.ts
@@ -3,11 +3,17 @@
 
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import path from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
     plugins: [react()],
     base: '/coinage/',
+    resolve: {
+        alias: {
+            '@': path.resolve(__dirname, './src'),
+        },
+    },
     test: {
         globals: true,
         environment: 'jsdom',
@@ -15,5 +21,8 @@ export default defineConfig({
         // you might want to disable it, if you don't have tests that rely on CSS
         // since parsing CSS is slow
         css: true,
+        alias: {
+            '@': path.resolve(__dirname, './src'),
+        },
     },
 });
